Add files via upload
In this analysis, global pairwise sequence alignment is performed to determine the similarity between human PPAR alpha and mouse PPAR alpha. To visualize  a comparison of the ligand binding domain residues between the two sequences, the method sequence logos was used. In part 2 of the analysis homology modeling was created using a target mouse PPAR alpha sequence and a known structure of human PPAR alpha ligand binding domain as the template. PyMOL was used to visually analyze the differences between the homology model and the human PPAR alpha crystal structure.